more on how Rx really works
It's a bit more safe how we handle messages now. Not still perfect. Before, we were Acking after OnNext but that was quite bad. Now, I removed that and it's not Acking at all now. Depening on the concurrency level, it will only pull certain amount of messages and will stop after that as non of the fetched messages are Acked. Next pr will make that happen.